Update controller to check database

+++ app/Http/Controllers/MangaController.php
@@ -4,8 +4,6 @@ namespace App\Http\Controllers;
 
 use App\Models\Manga;
 use Illuminate\Http\Request;
-use Illuminate\Support\Facades\DB;
-use Illuminate\Support\Facades\Http;
 
 class MangaController extends Controller
 {
@@ -16,7 +14,7 @@ class MangaController extends Controller
     public function index()
     {
         //Fetch all manga from database
-        $mangas = DB::table('mangas')->orderBy('created_at', 'ASC')->get();
+        $mangas = Manga::orderBy('created_at', 'ASC')->get();
         return view('welcome', ['mangas' => $mangas]);
     }
 
@@ -37,17 +35,35 @@ class MangaController extends Controller
     */
     public function create(Request $request)
     {
-        //dd($request->mal_id);
-        Manga::create([
-            'mal_id' => (int) $request->mal_id,
-            'eng_title' => $request->eng_title,
-            'jp_title' => $request->jp_title,
-            'author' => $request->author,
-            'run_start' => $request->run_start,
-            'run_end' => $request->run_end,
-            'status' => $request->status
-        ]);
-        return redirect('/')->with('status', 'Successfully added manga!');
+        $message = "";
+        //Check if manga is already added in list
+        if (Manga::where('mal_id', (int) $request->mal_id)->exists())
+        {
+            $message = "Manga already in list";
+        }
+        else
+        {
+            //Attempt to create record
+            try
+            {
+                //Use create method to add to database using request information.
+                Manga::create([
+                    'mal_id' => (int) $request->mal_id,
+                    'eng_title' => $request->eng_title,
+                    'jp_title' => $request->jp_title,
+                    'author' => $request->author,
+                    'run_start' => $request->run_start,
+                    'run_end' => $request->run_end,
+                    'status' => $request->status
+                ]);
+            } 
+            catch (Exception $e)
+            {
+                $message =  "Faied to add manga!";
+            }
+        }
+        
+        return redirect('/')->with('status', $message);
     }
 
     public function update(Request $request)
@@ -55,14 +71,14 @@ class MangaController extends Controller
         $message = "";
         //Check if record exists in database
         //https://devdojo.com/bobbyiliev/how-to-check-if-a-record-exists-with-laravel-eloquent
-        if (Manga::where('record_id', $request->record_id)->exists())
+        if (Manga::where('id', $request->id)->exists())
         {
             //If record exists
 
             try
             {
                 //Get record from database.
-                $manga = Manga::find($request->record_id);
+                $manga = Manga::find($request->id);
 
                 //Change status
                 $manga->status = $request->status;
@@ -91,10 +107,33 @@ class MangaController extends Controller
 
     public function destroy(Request $request)
     {
-        $manga = Manga::find($request->record_id);
+        $message = "";
+        //Check if record exists in database
+        //https://devdojo.com/bobbyiliev/how-to-check-if-a-record-exists-with-laravel-eloquent
+        if (Manga::where('id', $request->id)->exists())
+        {
+            //Attempt to retrieve record and delete it. 
+            try
+            {
+                //Retrieve record
+                $manga = Manga::find($request->id);
 
-        $manga->delete();
+                //Delete record
+                $manga->delete();
 
-        return redirect('/')->with('status', 'Successfully removed manga!');
+                $message .= 'Successfully removed manga!';
+            }
+            catch (Exception $e)
+            {
+                $message .= "Failed to delete manga!";
+            }
+        }
+        else
+        {
+            $message .= "Manga does not exist in list to remove.";
+        }
+        
+
+        return redirect('/')->with('status', $message);
     }
 }
